Workaround for issue #6
This commit applies a stop-gap workaround for the IInput-related issues
described in issue #6. It just applies to `jrte.Jrte.main()` the same
logic as `test.FileRunner.main()` -- the input file path is specified as
an input argument and the entire file contents are read into RAM. This
workaround just obviates inclusion of jrte-HEAD-test.jar in classpath.

Tests have been extended to include runs with FileRunner.main() for
benchmarking and Jrte.main() as for regular runs. Output equivalence
checks for equivalent output from verbose gc interval times extraction
via two different FSTs and one regex.

The `etc/sh/jrte.sh` script demonstrates how to use jrte to transduce a
file.
```
java -cp jrte-HEAD.jar com.characterforming.jrte.Jrte \
[--nil] <transducer-name> <input-filepath> <gearbox-filepath>
```
The `--nil` option presents a `!nil` signal to the transduction before
presenting the file contents.

This is all that I intend to do with this issue #6 for now as a good fix
will require replacing the entire IInput framework and transducing from
raw byte streams; this will be undertaken in connection with issue #15.

Signed-off-by: jrte <jrte.project@gmail.com>