WIP: Don't try this at home
Looks like another Big Rewrite is coming...

1) IInput and the input stack need to be reworked because current design
forces creation of innumerable small objects to wrap input byte arrays.
The stack should be composed of an array of clips

   (byte[], position, limit)

that reference a byte array and define a window of bytes to scan. These
clips would be statically bound to the input stack. The input stack size
would be fixed or could grow up to a fixed limit.

2) The run() loop burns a lot of oil. If I recall from years ago I had
to roll out the effector selection loop as for run() implementation at
bottom of git history pile for Transduciton.java.

3) The CounterTest and PasteCountTest transductions blow the heap out
to 100s of MB, they should run in 12 MB. The while design needs to be
reworked to eliminate unnecessary proliferation of small wrapper
objects.

4) The Transduction target effectors need to be inlined for invoke(int)
as well as for invoke(), using bound effector getParameter(int) to
select parameter instance.

Otherwise, all tests pass. Commiting this as a checkpoint as I start to
tear it apart again.

Signed-off-by: jrte <jrte.project@gmail.com>