WIP: Fixup Java regex used in test runoffs
The java regex for pulling fields from linux kernel logs was not
matching anything so was competing much faster than expected. This is
fixed and the performance runoff metrics look a lot more like they do
in the graphic posted in ribose/README.md. The outputs from the
LinuxKernel variant transducers and equivalent java regex run are
checked for identical content in the test runs and the build will
fail if all 4 files are not identical.

Also starting some work on API documentation. Removed javadoc task from
build.xml and removed default build task dependency on javadoc. Ant
build file for javadoc is now in javadoc.xml. API is a WIP, incompplete
and volatile.

Signed-off-by: jrte <jrte.project@gmail.com>