Add GC regression logging to build
- for file equivalence test runs
- list heap reclaimed, retained and total sizes after each G1 pause
- append results to running regression.gc (checked in with each build)
- list transduction throughput
- script in etc/sh/regress
- regression.gc in this commit are from stable precommit state

Legend for regression.gc:
1. SHA for current ribose commit
2. SHA for most recent commit containing etc/sh/regression
3. timestamp - transducer(.gc)
4. reclaimed, retained, total (labels)
5. reclaimed, retained, total (volume, per gc, in megabytes)

The regex GC stats are confounded by a 20mb heap charge for the UTF-8
raw bytes that must be preloaded and another 40mb charge for the
decoded char array. Java regex cannot accept segmented input AFAIK. The
relevant GC metric for comparison with risose is the reclaimed metric.

The ribose throughput stats are confounded by blocking I/O (input and
output). Regex input in preloaded into the heap but output is blocking
and there is encoder overhead on output as well, which is not a factor in
ribose runs.

TODO: throughput regression logging (regression.thruput)