WIP #30: Add mproduct instrumentation
The linux kernel transductions benefit a bit but there is regression in
the verbose gc transductions. Can't account for the latter but these
transductions spend almost all the time failing a short mproduct run
before immediately switching to a long msum run.

See #30 for benchmark results (compare with #31 benchmarks).

- first cut instruments transducers with mproduct effector
- mproduct effector invoked on transition into first sequential input
- mproduct trap implemented in Transductor
- FileRunner metrics in build show mproduct effector invocations/kb

TODO:
- eliminate redundant states after mproduct instrumentation
- coalesce input symbols after eliminating redundant states