Update ginr executables version 2.1.0f->2.1.0g
This commit started as a spell checking expedition but a ginr
bug popped up and additional changes were required. The
functional changes to ribose code are limited to Model,java,
which now forces model compilation to fail if no transducers
are compiled into the model. It has some updates suggested
by Sonar as well and more Sonar-related changes will be
forthcoming in the next series of commits.

There are some substantive edits in README.md and the
javadoc comments.

The patterns script has been changed to provide more helpful
output relating to ginr compilation errors.

Signed-off-by: jrte <jrte.project@gmail.com>